voteComment() function implemented, Description detail
If a user upVoted a comment, it is restricted him to upvote again.
Same for downVoting.
But If user upVoted recently and changes it to downVote, it is not
restricted. upVote counts back and downVote is made after it (as a
result , rate is rate-2)
Same for the other way around. Result, rate is rate+2