FollowedTopics functions implemented. Details description
followTopic takes a topicId, returns true if it can follow the topic,
returns false if it is followed already.

unfollowTopic takes a topicId, returns true if it can unfollow the
followed topic, returns false if the topic is not followed already.

checkFollowedTopic, takes a topicId, returns if user is following it.

getFollowedUserTopics, returns List<Topics> of the user.